Kick off async process on github posts. Add command for easier testing.
