+++ ployst/__init__.py
@@ -1,0 +1,3 @@
+from __future__ import absolute_import
+
+from .celery import app as celery_app  # NOQA
