+++ ployst/settings/test.py
@@ -1,0 +1,3 @@
+from .dev import *  # NOQA
+
+CELERY_ALWAYS_EAGER = True
