+++ requirements/runtime.txt
@@ -14,3 +14,5 @@ apitopy==0.3.0
 
 # Git provider requirements
 GitPython==0.3.2.RC1             # access git repositories
+
+celery==3.1.9
